+++ packages/components/.storybook/decorators/withProviders.tsx
@@ -1,4 +1,4 @@
-import React from 'react'
+import React, { useEffect } from 'react'
 
 import { LoadingState, ThemeProvider } from '@baseapp-frontend/design-system'
 import { RelayTestProvider, createTestEnvironment } from '@baseapp-frontend/graphql'
@@ -17,13 +17,21 @@ const withProviders = (Story: StoryFn, context: StoryContext) => {
   // TODO: registering a few tailwind classess (used by @baseapp-frontend/design-system components), need to figure out why the @baseapp-frontend/components storybook are not including it correctly
   // pb-3 px-3 w-full rounded-md bg-background-neutral px-2 py-1
   const relayMockEnvironment = createTestEnvironment()
-  const { environment, queueOperationResolver } = relayMockEnvironment
+  const { environment, queueOperationResolver, resolveMostRecentOperation } = relayMockEnvironment
 
   context.parameters.relayMockEnvironment = relayMockEnvironment
 
-  const mockResolvers = context.parameters.mockResolvers || {}
+  const mockResolvers = context.parameters.mockResolvers || undefined
+  const mockData = context.parameters.mockData || {}
 
-  queueOperationResolver(mockResolvers)
+  useEffect(() => {
+    if (!!mockResolvers) {
+      queueOperationResolver(mockResolvers)
+    }
+    if (mockData) {
+      resolveMostRecentOperation({ mockResolvers, data: mockData })
+    }
+  }, [mockData, resolveMostRecentOperation, queueOperationResolver, mockResolvers])
 
   return (
     <QueryClientProvider client={queryClient}>