+++ packages/wagtail/modules/providers/index.ts
@@ -1,0 +1,4 @@
+export { useWagtailPagesContext } from './WagtailPagesProvider/context'
+
+export { PageLayoutProvider } from './PageLayoutProvider'
+export { usePageLayoutContext } from './PageLayoutProvider/context'