+++ packages/wagtail/modules/services/Wagtail/index.ts
@@ -1,0 +1,7 @@
+import axios from '../axios'
+
+export class WagtailAPI {
+  static baseURL = `${process.env.NEXT_PUBLIC_WAGTAIL_API_BASE_URL}`
+
+  static axios = axios
+}