+++ packages/wagtail/modules/services/index.ts
@@ -1,0 +1,2 @@
+export type * from './Wagtail/PagesAPI/types'
+export type * from './Wagtail/types'