+++ packages/design-system/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @baseapp-frontend/design-system
 
+## 0.0.20
+
+### Patch Changes
+
+- Create `Searchbar` component. 
+- Add `NoMessagesIcon` icon. 
+- Export `PureTextFieldProps` type. 
+
+## 0.0.19
+
+### Patch Changes
+
+- Added the Searchbar component
+
+## 0.0.18
+
+### Patch Changes
+
+- Impose file size limits on Dropzones
+
+## 0.0.17
+
+### Patch Changes
+
+- Storybook cleanup
+
 ## 0.0.16
 
 ### Patch Changes